# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DisClone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers/{serverId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List server members
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /servers/{serverId}/members/{userId}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update member (nickname, roles)
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: kickMember
      tags: [members]
      summary: Kick a member
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kicked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /servers/{serverId}/bans/{userId}:
    post:
      operationId: banMember
      tags: [members]
      summary: Ban a member
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: unbanMember
      tags: [members]
      summary: Unban a member
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ROLES ───────────────────────────────────────────────────────────────────
...
  /friends/add:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "200":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/{userId}/accept:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a friend request
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/{userId}/decline:
    post:
      operationId: declineFriendRequest
      tags: [friends]
      summary: Decline a friend request
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/{userId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/{userId}/block:
    post:
      operationId: blockUser
      tags: [friends]
      summary: Block a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── INVITES ─────────────────────────────────────────────────────────────────
  /servers/{serverId}/invites:
    get:
      operationId: listInvites
      tags: [invites]
      summary: List invites for a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
    post:
      operationId: createInvite
      tags: [invites]
      summary: Create an invite
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Created invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /invites/{code}:
    get:
      operationId: getInvite
      tags: [invites]
      summary: Get invite details by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /invites/{code}/use:
    post:
      operationId: useInvite
      tags: [invites]
      summary: Join server via invite code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"

  # ─── ACTIVITY ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, username, discriminator, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        discriminator:
          type: string
        displayName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        accentColor:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, idle, dnd, invisible, offline]
        customStatus:
          type: ["string", "null"]
        customStatusEmoji:
          type: ["string", "null"]
        isNitro:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── SERVER ───────────────────────────────────────────────────────────────
    Server:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        ownerId:
          type: integer
        isPublic:
          type: boolean
        memberCount:
          type: integer
        boostLevel:
          type: integer
        boostCount:
          type: integer
        vanityUrl:
          type: ["string", "null"]
        verificationLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Member:
      type: object
      required: [userId, serverId, joinedAt]
      properties:
        userId:
          type: integer
        serverId:
          type: integer
        nickname:
          type: ["string", "null"]
        roles:
          type: array
          items:
            type: integer
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          type: string
          format: date-time
    MemberUpdate:
      type: object
      properties:
        nickname:
          type: ["string", "null"]
        roles:
          type: array
          items:
            type: integer

    # ─── ROLE ─────────────────────────────────────────────────────────────────
...
    FriendRequestInput:
      type: object
      required: [username]
      properties:
        username:
          type: string

    # ─── INVITE ───────────────────────────────────────────────────────────────
    Invite:
      type: object
      required: [code, serverId, inviterId, createdAt]
      properties:
        code:
          type: string
        serverId:
          type: integer
        server:
          $ref: "#/components/schemas/Server"
        channelId:
          type: ["integer", "null"]
        inviterId:
          type: integer
        inviter:
          $ref: "#/components/schemas/User"
        uses:
          type: integer
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    InviteInput:
      type: object
      properties:
        channelId:
          type: integer
        maxUses:
          type: integer
        expiresInHours:
          type: integer

    # ─── ACTIVITY ─────────────────────────────────────────────────────────────
...