# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisClone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  /** @nullable */
  nickname?: string | null;
  roles?: number[];
}