# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DisClone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Server {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  ownerId: number;
  isPublic?: boolean;
  memberCount?: number;
  boostLevel?: number;
  boostCount?: number;
  /** @nullable */
  vanityUrl?: string | null;
  verificationLevel?: number;
  createdAt: Date;
}